Refactor: repackage service package
