+++ src/main/java/com/project/Project/auth/provider/JwtProvider.java
@@ -6,7 +6,7 @@ import com.project.Project.auth.dto.Token;
 import com.project.Project.auth.exception.JwtException;
 import com.project.Project.auth.service.TokenService;
 import com.project.Project.domain.Member;
-import com.project.Project.service.MemberService;
+import com.project.Project.service.member.MemberService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.authentication.AuthenticationProvider;
