+++ src/main/java/com/project/Project/auth/service/TokenService.java
@@ -2,7 +2,7 @@ package com.project.Project.auth.service;
 
 import com.project.Project.auth.dto.Token;
 import com.project.Project.domain.Member;
-import com.project.Project.service.MemberService;
+import com.project.Project.service.member.MemberService;
 import io.jsonwebtoken.*;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
