+++ src/main/java/com/project/Project/controller/building/BuildingRestController.java
@@ -8,8 +8,8 @@ import com.project.Project.controller.building.dto.BuildingResponseDto;
 import com.project.Project.domain.building.Building;
 import com.project.Project.repository.projection.building.OnlyBuildingIdAndCoord;
 import com.project.Project.serializer.building.BuildingSerializer;
-import com.project.Project.service.BuildingService;
-import com.project.Project.service.ReviewService;
+import com.project.Project.service.building.BuildingService;
+import com.project.Project.service.review.ReviewService;
 import com.project.Project.validator.ExistBuilding;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
