+++ src/main/java/com/project/Project/controller/review/controller/ReviewRestController.java
@@ -13,10 +13,10 @@ import com.project.Project.domain.enums.KeywordEnum;
 import com.project.Project.domain.review.Review;
 import com.project.Project.domain.room.Room;
 import com.project.Project.serializer.review.ReviewSerializer;
-import com.project.Project.service.BuildingService;
-import com.project.Project.service.ReviewImageService;
-import com.project.Project.service.ReviewService;
-import com.project.Project.service.RoomService;
+import com.project.Project.service.building.BuildingService;
+import com.project.Project.service.review.ReviewImageService;
+import com.project.Project.service.review.ReviewService;
+import com.project.Project.service.room.RoomService;
 import com.project.Project.validator.ExistBuilding;
 import com.project.Project.validator.ExistReview;
 import com.project.Project.validator.ExistRoom;
