+++ src/main/java/com/project/Project/service/FileProcessService.java
@@ -1,4 +1,4 @@
-package com.project.Project.service.impl;
+package com.project.Project.service;
 
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.project.Project.domain.enums.FileFolder;
@@ -46,6 +46,6 @@ public class FileProcessService {
 
     private String getFileName(String url) {
         String[] paths = url.split("/");
-        return paths[paths.length-2] + "/" + paths[paths.length-1];
+        return paths[paths.length - 2] + "/" + paths[paths.length - 1];
     }
 }
