+++ src/main/java/com/project/Project/service/building/BuildingService.java
@@ -1,4 +1,4 @@
-package com.project.Project.service;
+package com.project.Project.service.building;
 
 import com.project.Project.controller.building.dto.BuildingOptionalDto;
 import com.project.Project.domain.building.Building;
@@ -12,13 +12,19 @@ import java.util.Optional;
 
 public interface BuildingService {
     List<Building> getBuildingListByBuildingIds(List<Long> buildingIds, Long cursorId, Pageable page);
+
     Building getBuildingByBuildingId(Long buildingId);
+
     List<Building> getBuildingsBySearch(String params, Long cursorId, Pageable page);
+
     Optional<Building> findByAddress(Address address);
+
     List<OnlyBuildingIdAndCoord> getAllBuildingsIdAndCoord();
+
     Building createBuilding(Address address, BuildingOptionalDto buildingOptionalDto);
 
     Building createBuilding(Address address, Coordinate coordinate, BuildingOptionalDto buildingOptionalDto);
+
     List<Building> createBuilding(String address);
 
 }
