+++ src/main/java/com/project/Project/service/review/ReviewCategoryService.java
@@ -1,4 +1,4 @@
-package com.project.Project.service;
+package com.project.Project.service.review;
 
 import com.project.Project.domain.enums.ReviewCategoryEnum;
 import com.project.Project.domain.review.ReviewCategory;
