+++ src/main/java/com/project/Project/service/review/ReviewImageService.java
@@ -1,4 +1,4 @@
-package com.project.Project.service;
+package com.project.Project.service.review;
 
 import com.project.Project.domain.review.Review;
 import org.springframework.web.multipart.MultipartFile;
