+++ src/main/java/com/project/Project/service/review/ReviewKeywordService.java
@@ -1,4 +1,4 @@
-package com.project.Project.service;
+package com.project.Project.service.review;
 
 import com.project.Project.domain.enums.DTypeEnum;
 import com.project.Project.domain.enums.KeywordEnum;
