+++ src/main/java/com/project/Project/service/review/ReviewService.java
@@ -1,4 +1,4 @@
-package com.project.Project.service;
+package com.project.Project.service.review;
 
 import com.project.Project.domain.building.Building;
 import com.project.Project.domain.review.Review;
@@ -9,6 +9,7 @@ import java.util.List;
 
 public interface ReviewService {
     List<Review> getReviewListByBuildingId(Long buildingId, Long cursorId, Pageable page);
+
     List<Review> getReviewListByRoomId(Long roomId, Long cursorId, Pageable page);
 
     Long deleteById(Long reviewId);
