+++ src/main/java/com/project/Project/service/review/impl/ReviewImageServiceImpl.java
@@ -1,10 +1,11 @@
-package com.project.Project.service.impl;
+package com.project.Project.service.review.impl;
 
 import com.project.Project.domain.enums.FileFolder;
 import com.project.Project.domain.review.Review;
 import com.project.Project.domain.review.ReviewImage;
 import com.project.Project.repository.review.ReviewImageRepository;
-import com.project.Project.service.ReviewImageService;
+import com.project.Project.service.FileProcessService;
+import com.project.Project.service.review.ReviewImageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
