+++ src/main/java/com/project/Project/service/room/RoomService.java
@@ -1,4 +1,4 @@
-package com.project.Project.service;
+package com.project.Project.service.room;
 
 import com.project.Project.domain.building.Building;
 import com.project.Project.domain.room.Room;
@@ -7,5 +7,6 @@ import java.util.Optional;
 
 public interface RoomService {
     Room createRoom(Building building, Integer lineNumber, Integer roomNumber);
+
     Optional<Room> findByBuildingAndLineNumberAndRoomNumber(Building building, Integer lineNumber, Integer roomNumber);
 }
