+++ src/test/java/com/project/Project/repository/ReviewKeywordRepoTest.java
@@ -4,8 +4,8 @@ import com.project.Project.domain.enums.DTypeEnum;
 import com.project.Project.domain.enums.KeywordEnum;
 import com.project.Project.domain.review.ReviewKeyword;
 import com.project.Project.repository.review.ReviewKeywordRepository;
-import com.project.Project.service.ReviewCategoryService;
-import com.project.Project.service.ReviewKeywordService;
+import com.project.Project.service.review.ReviewCategoryService;
+import com.project.Project.service.review.ReviewKeywordService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
